requirements.txt
bs4==0.0.2
lxml==5.3.0
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1